Extract complicated methods from application_controller to modules and/or better places for for them
